src/UI/requirements.txt
base58==2.1.1
certifi==2022.6.15.1
charset-normalizer==2.1.1
contourpy==1.0.6
cycler==0.11.0
ecdsa==0.18.0
fonttools==4.38.0
future==0.18.2
hdwallet==2.1.1
idna==3.3
kiwisolver==1.4.4
matplotlib==3.6.1
mnemonic==0.20
numexpr==2.8.4
numpy==1.23.4
packaging==21.3
pandas==1.5.1
pandastable==0.13.0
Pillow==9.3.0
pycryptodome==3.15.0
pyparsing==3.0.9
pysha3==1.0.2
python-dateutil==2.8.2
pytz==2022.5
requests==2.28.1
six==1.16.0
urllib3==1.26.12
xlrd==2.0.1
